+++ server/world.go
@@ -5,7 +5,7 @@ type world struct {
 }
 
 // Returns nil if unit doesn't exist
-func (w *world) Unit(id int) IUnit {
+func (w *world) Unit(id int64) IUnit {
 	index := w.unitIndex(id)
 	if index < 0 {
 		return nil
@@ -17,7 +17,7 @@ func (w *world) Units() []IUnit {
 	return w.units
 }
 
-func (w *world) unitIndex(id int) int {
+func (w *world) unitIndex(id int64) int {
 	for index, unit := range w.units {
 		if unit.Id() == id {
 			return index
@@ -30,7 +30,7 @@ func (w *world) AddUnit(unit IUnit) {
 	w.units = append(w.units, unit)
 }
 
-func (w *world) RemoveUnit(id int) (success bool) {
+func (w *world) RemoveUnit(id int64) (success bool) {
 	index := w.unitIndex(id)
 	if index < 0 {
 		return false
